Innerloop wf staging

+++ tap-tests/e2e/accelerator.go
@@ -20,11 +20,15 @@ func ListAccelerators() {
 	}
 }
 
-func GenerateAcceleratorProject(label string, projectName string, repositoryPrefix string, unzip bool) {
+func GenerateAcceleratorProject(label string, projectName string, repositoryPrefix string, unzip bool, serverIP string) {
 	log.Printf("Generating accelerator project: %s", label)
-	tap.RunWithBash(fmt.Sprintf(`tanzu accelerator generate %s --options '{"projectName":"%s", "repositoryPrefix":"%s", "includeKubernetes": true}'`, label, projectName, repositoryPrefix))
+	tap.RunWithBash(fmt.Sprintf(`tanzu accelerator generate %s --options '{"projectName":"%s", "repositoryPrefix":"%s", "includeKubernetes": true}' --server-url http://%s`, label, projectName, repositoryPrefix, serverIP))
 	if unzip {
-		tap.Run(fmt.Sprintf("rm -r %s", projectName))
+		out, err := tap.Run_AllowError(fmt.Sprintf("ls -lt %s", projectName))
+		if err == nil{
+			tap.Run(fmt.Sprintf("rm -r %s", projectName))
+		}
+		log.Printf("Output: \n%s", string(out))
 		tap.Run(fmt.Sprintf("unzip %s.zip", projectName))
 	}
 }
